% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.analysis.list}
\alias{gx.analysis.list}
\title{Lists available analysis tools}
\usage{
gx.analysis.list()
}
\value{
a list containing available analysis tools and their tool group
}
\description{
Returns a list containing the available analysis tools
and their tool group. Tools are organized into thematic groups
like \emph{Data manipulation} or \emph{Site analysis}.
}
\seealso{
\code{\link{gx.analysis}}

\code{\link{gx.analysis.parameters}}

\code{\link[rbiouml]{biouml.analysis.list}}
}
\keyword{analysis,}
\keyword{list}
