% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.analysis.parameters}
\alias{gx.analysis.parameters}
\title{Shows parameters for the specified analysis tool}
\usage{
gx.analysis.parameters(analysisName)
}
\description{
Wraps rbiouml function biouml.analysis.parameters
}
\examples{
gx.analysis.parameters()

}
\seealso{
\code{\link{rbiouml::biouml.analysis.parameters}}
}
\keyword{analysis,}
\keyword{parameters}
