% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.classifyChipSeqTargets}
\alias{gx.classifyChipSeqTargets}
\title{Runs the workflow \emph{ChIP-Seq - Identify and classify target genes}}
\usage{
gx.classifyChipSeqTargets(inputTrack, species = "Human (Homo sapiens)",
  resultFolder, skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTrack}{path to track with input ChIP-seq intervals}

\item{species}{species of the input track genome}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\value{
the job id of the submitted task. The job id can be used to retrieve information about the status of the analysis.
}
\description{
Runs the workflow \emph{ChIP-Seq - Identify and classify target genes}
}
\keyword{ChIP-seq,}
\keyword{classification,}
\keyword{gene}
\keyword{target}
\keyword{workflow,}
