% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.createFolder}
\alias{gx.createFolder}
\title{Create a folder}
\usage{
gx.createFolder(path, folderName)
}
\arguments{
\item{path}{parent folder in which the new folder is created}

\item{folderName}{name of the new folder}
}
\value{
a string containing the status of the request in JSON format
}
\description{
This function creates a new folder within an existing parent folder.
}
