% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.createProject}
\alias{gx.createProject}
\title{Create a new project}
\usage{
gx.createProject(name, description = "New platform project")
}
\arguments{
\item{name}{name of the project}

\item{description}{a short description of the project}
}
\description{
Create a new project
}
\examples{

# Note that the project name probably requires adaptation

gx.createProject("example-project","example description")

}
