% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.delete}
\alias{gx.delete}
\title{Delete a workspace item}
\usage{
gx.delete(folder, name)
}
\arguments{
\item{folder}{folder that contains the item}

\item{name}{name of the item}
}
\value{
type and value of the response. A type of 0 indicates success.
}
\description{
This function deletes an item within the workspace. The item may be
a folder or another type of element. 
Handle with care, because deletion may be irreversible.
}
