% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.ebarraysWorkflow}
\alias{gx.ebarraysWorkflow}
\title{Runs the workflow \emph{Compute differentially expressed genes using EBarrays}}
\usage{
gx.ebarraysWorkflow(inputTable, probeType = "Probes: Affymetrix",
  species = "Human (Homo sapiens)", controlName = "Control",
  controlColumns = c(), conditions = list(), resultFolder,
  skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with expression data}

\item{probeType}{type of probes}

\item{species}{species of the input track genome}

\item{controlName}{name for the control column group}

\item{controlColumns}{list of control column names}

\item{conditions}{a list with condition names and columns}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
This function computes differentially expressed genes using EBarrays.
}
\details{
A list of column names that make up the control condition is necessary. Furthermore,
there must be at least one more condition in \emph{conditions} list parameter. At most
four conditions (besides the control) can be considered.
Their names are specified within the list like \code{group_2="name of second group", ..., group_5=...}.
Note that there is no \emph{group_1}. Furthermore, one needs to assign columns of the input table to conditions (groups).
This information is provided within the same parameter list like
\code{Columns_group_2=c("first column name for group_2",...), Columns_group_3=c(...)}.
}
\keyword{EBarrays}
\keyword{differential}
\keyword{expression,}
\keyword{workflow,}
