% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.exists}
\alias{gx.exists}
\title{Finds out if a function exists}
\usage{
gx.exists(name)
}
\arguments{
\item{name}{name of a possible function}
}
\description{
This function finds out whether a function exists within the geneXplainR package or not.
}
\keyword{exist,}
\keyword{function}
