% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.explainMyGenes}
\alias{gx.explainMyGenes}
\title{Runs the workflow \emph{Explain my genes}}
\usage{
gx.explainMyGenes(inputTable, species = "Human (Homo sapiens)",
  resultFolder, skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with gene ids}

\item{species}{species of the input track genome}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\value{
the job id of the submitted task. The job id can be used to retrieve information about the status of the analysis.
}
\description{
Runs the workflow \emph{Explain my genes}
}
\keyword{analysis}
\keyword{upstream}
\keyword{workflow,}
