% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.export.parameters}
\alias{gx.export.parameters}
\title{Returns parameters defined for given export and path to export from}
\usage{
gx.export.parameters(path, exporter)
}
\arguments{
\item{path}{path of the object to export}

\item{exporter}{exporter to use for the export}
}
\description{
Wraps rbiouml function biouml.export.parameters
}
\seealso{
\code{\link[rbiouml]{biouml.export.parameters}}
}
\keyword{exporter,}
\keyword{parameters}
