% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.export}
\alias{gx.export}
\title{Exports item using specified exporter}
\usage{
gx.export(path, exporter = "Tab-separated text (*.txt)",
  exporter.params = list(), target.file = "genexplain.out")
}
\arguments{
\item{path}{path to export}

\item{exporter}{exporter to use for export}

\item{exporter.params}{parameters of the exporter}

\item{target.file}{local file to export to}
}
\description{
Wraps rbiouml function biouml.export
}
\seealso{
\code{\link[rbiouml]{biouml.export}}
}
\keyword{export}
