% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.focusedUpstreamAnalysis}
\alias{gx.focusedUpstreamAnalysis}
\title{Runs the workflow \emph{Focused upstream analysis (TRANSFAC(R) and TRANSPATH(R))}}
\usage{
gx.focusedUpstreamAnalysis(inputYesSet, inputNoSet,
  species = "Human (Homo sapiens)", profile, foldEnriched = 1,
  promoterStart = -1000, promoterEnd = 100, resultFolder,
  allowBigInput = F, skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputYesSet}{yes/positive/foreground gene set}

\item{inputNoSet}{no/negative/background gene set}

\item{species}{species of the input track genome}

\item{profile}{matrix profile}

\item{foldEnriched}{filter by fold enrichment}

\item{promoterStart}{first base of a promoter relative to TSS}

\item{promoterEnd}{last base of a promoter relative to TSS}

\item{resultFolder}{path of result folder}

\item{allowBigInput}{set true to allow for large data sets}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
Runs the workflow \emph{Focused upstream analysis (TRANSFAC(R) and TRANSPATH(R))}
}
\keyword{GeneWays}
\keyword{TRANSFAC,}
\keyword{analysis,}
\keyword{upstream}
\keyword{workflow,}
