% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.get}
\alias{gx.get}
\title{Gets a table from the platform workspace}
\usage{
gx.get(path)
}
\arguments{
\item{path}{path of object to load into a data.frame}
}
\value{
a list containing the table
}
\description{
Returns a list containing the specified table.
}
\seealso{
\code{\link[rbiouml]{biouml.get}}
}
\keyword{get}
