% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.import.parameters}
\alias{gx.import.parameters}
\title{Returns parameters defined for given importer and path to import to}
\usage{
gx.import.parameters(path, importer)
}
\arguments{
\item{path}{path to import to}

\item{importer}{importer whose parameters will be shown}
}
\description{
Wraps rbiouml function biouml.import.parameters
}
\seealso{
\code{\link[rbiouml]{biouml.import.parameters}}
}
\keyword{import,}
\keyword{parameters}
