% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.import}
\alias{gx.import}
\title{Imports a file into the platform}
\usage{
gx.import(file, parentPath, importer, importer.params = list())
}
\description{
Wraps rbiouml function biouml.import
}
\examples{
gx.import("example/file/path", "example/platform/path", "example-importer", list())

}
\seealso{
\code{\link{rbiouml::biouml.import}}
}
\keyword{import}
