% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.importBedFile}
\alias{gx.importBedFile}
\title{Import a BED file}
\usage{
gx.importBedFile(bedFile, destPath,
  genomeDb = "Ensembl 91.38 Human (hg38)", genomeId = "hg38",
  wait = T, verbose = T)
}
\arguments{
\item{bedFile}{BED file to import}

\item{destPath}{Platform path of resulting track item}

\item{genomeDb}{Genome database}

\item{genomeId}{Genome id string, e.g. mm10, hg38}

\item{wait}{True to wait for job to complete}

\item{verbose}{True for more progress info}
}
\value{
a string containing the status of the request in JSON format
}
\description{
Imports a BED file to specified data item
}
