% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.importers}
\alias{gx.importers}
\title{Lists available importers}
\usage{
gx.importers()
}
\description{
Wraps rbiouml function biouml.importers
}
\seealso{
\code{\link[rbiouml]{biouml.importers}}
}
\keyword{importer,}
\keyword{list}
