% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.importTable}
\alias{gx.importTable}
\title{Import a table from a text file}
\usage{
gx.importTable(tableFile, platformFolder, tableName = "imported_table",
  processQuotes = TRUE, delim = "Tab", headerRow = 1, dataRow = 2,
  commentString = "", columnForID = "", addSuffix = FALSE,
  tableType = "", species = "Unspecified")
}
\arguments{
\item{tableFile}{path to file with the table}

\item{platformFolder}{folder into which the table is imported}

\item{tableName}{name for the table within the platform}

\item{processQuotes}{true to process quotation marks}

\item{delim}{column delimiter, one of the strings (Tab, Spaces, Commas, Pipes)}

\item{headerRow}{row index of the table header}

\item{dataRow}{row index of the first data row}

\item{commentString}{string to recognize comment lines}

\item{columnForID}{name of column with ids}

\item{addSuffix}{true to add suffix to ensure unique ids}

\item{tableType}{type of platform table, e.g. Genes: Ensembl}

\item{species}{Latin name of species}
}
\value{
a string containing the status of the request in JSON format
}
\description{
This function imports a table into an existing platform folder.
}
