% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.isElement}
\alias{gx.isElement}
\title{Finds out if an item is part of a folder}
\usage{
gx.isElement(path, name)
}
\arguments{
\item{path}{platform path}

\item{name}{name of the item}
}
\description{
This function checks whether an item exists in a certain path.
}
\keyword{exist,}
\keyword{isElement,}
\keyword{item,}
\keyword{path}
