% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.job.info}
\alias{gx.job.info}
\title{Gets status info for a job running on the platform}
\usage{
gx.job.info(jobID)
}
\arguments{
\item{jobID}{id of the platform task}
}
\value{
a list with status info, response type of the job, the requested result and the analysis log
}
\description{
This function retrieves status information for a task running
on the platform. A common use case is to check the status of
tasks invoked asynchronously using \emph{gx.analysis} and \emph{gx.workflow} with
\code{wait=F}.
}
\seealso{
\code{\link[rbiouml]{biouml.job.info}}
}
\keyword{info}
\keyword{job,}
