% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.limmaWorkflow}
\alias{gx.limmaWorkflow}
\title{Runs the workflow \emph{Compute differentially expressed genes using Limma}}
\usage{
gx.limmaWorkflow(inputTable, probeType = "Probes: Affymetrix",
  species = "Human (Homo sapiens)", conditions = list(), resultFolder,
  skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with expression data}

\item{probeType}{type of probes}

\item{species}{species of the input track genome}

\item{conditions}{a list with condition names and columns}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\value{
the job id of the submitted task. The job id can be used to retrieve information about the status of the analysis.
}
\description{
This function computes differentially expressed genes using Limma. Limma
output tables (one for each condition pair) are located in the result folder
named after the compared conditions.
}
\details{
The conditions need to be specified in the corresponding list parameter.
There can be up to five different conditions. Their names are specified
within the list like \code{Condition_1="name of first condition", ..., Condition_5=...}.
Furthermore, one needs to assign columns of the input table to conditions.
This information is provided within the same parameter list like
\code{1_Columns=c("first column name for condition 1",...), 2_Columns=c(...)}.
}
\keyword{differential}
\keyword{expression,}
\keyword{limma}
\keyword{workflow,}
