% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.logout}
\alias{gx.logout}
\title{Terminates an existing platform session}
\usage{
gx.logout()
}
\value{
\code{NULL}
}
\description{
Signs out of a platform session.
}
\seealso{
\code{\link[rbiouml]{biouml.logout}}
}
\keyword{logout}
