% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.mapGenesToOntologies}
\alias{gx.mapGenesToOntologies}
\title{Runs the workflow \emph{Mapping to ontologies (Gene table)}}
\usage{
gx.mapGenesToOntologies(inputTable, species = "Human (Homo sapiens)",
  resultFolder, skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with gene ids}

\item{species}{species of the input track genome}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
Runs the workflow \emph{Mapping to ontologies (Gene table)}
}
\keyword{classification,}
\keyword{function}
\keyword{gene,}
\keyword{ontology,}
\keyword{workflow,}
