% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.put}
\alias{gx.put}
\title{Uploads a table to the platform}
\usage{
gx.put(path, value)
}
\description{
Wraps rbiouml function biouml.put
}
\examples{
gx.put("example/destination/path",data)

}
\seealso{
\code{\link{rbiouml::biouml.put}}
}
\keyword{put}
