% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.searchRegulators}
\alias{gx.searchRegulators}
\title{Regulator search}
\usage{
gx.searchRegulators(sourcePath, weightColumn, limitInputSize = FALSE,
  inputSizeLimit = 1000, maxRadius = 5, scoreCutoff = 0.2, bioHub,
  species = "Human (Homo sapiens)", calculateFdr = TRUE,
  fdrCutoff = 0.05, zscoreCutoff = 1, penalty = 0.1,
  isoformFactor = TRUE, contextDecorators = c(),
  removeNodeDecorators = c(), outputTable, wait = TRUE,
  verbose = TRUE)
}
\arguments{
\item{sourcePath}{Path of table with input protein/molecule set}

\item{weightColumn}{Name of column in input table with weights (optional)}

\item{limitInputSize}{True to take only the first N inputs}

\item{inputSizeLimit}{Number of inputs to use from input table}

\item{maxRadius}{Max. number of reactions/interactions to connect inputs to regulators}

\item{scoreCutoff}{Cutoff for regulator score}

\item{bioHub}{Molecular network database to use}

\item{species}{Species whose network shall be analyzed}

\item{calculateFdr}{True to create random inputs for Z-score calculation}

\item{fdrCutoff}{Cutoff to consider regulators significant}

\item{zscoreCutoff}{Cutoff for Z-score}

\item{penalty}{Penalty for non-input hits in regulator network}

\item{isoformFactor}{Normalize multi-forms}

\item{contextDecorators}{List of lists with context decorator parameters like list('A'=list(table='table path', column='column name', 'decay'=<value>), list(), ...)}

\item{removeNodeDecorators}{List of table paths specifying nodes to remove}

\item{outputTable}{Path to store analysis results}

\item{wait}{True to wait for job to complete}

\item{verbose}{True for more progress info}
}
\value{
a string containing the status of the request in JSON format
}
\description{
Searches for signal transduction regulators of input proteins
}
