% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.trackToGeneSet}
\alias{gx.trackToGeneSet}
\title{Track to gene set}
\usage{
gx.trackToGeneSet(tracks, species, from, to, resultTypes = c("Count"),
  allGenes = F, destPath, wait = T, verbose = T)
}
\arguments{
\item{tracks}{List of track paths}

\item{species}{Name of the species}

\item{from}{Gene region start relative to 5' end of Ensembl gene}

\item{to}{Gene region end relative to 3' end of Ensembl gene}

\item{resultTypes}{List of statistics to report (Schematic, + or -, Count, Count in exons, Count in introns, Count in 5', Count in 3', Structure, Positions)}

\item{allGenes}{True if all genes shall be reported regardless of hit}

\item{destPath}{output path}

\item{wait}{True to wait for job to complete}

\item{verbose}{True for more progress info}
}
\value{
a string containing the status of the request in JSON format
}
\description{
Maps one or more tracks to genes of the most recent Ensembl release
}
