% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.upstreamAnalysisTransfacGeneWays}
\alias{gx.upstreamAnalysisTransfacGeneWays}
\title{Runs the workflow \emph{Upstream analysis (TRANSFAC(R) and GeneWays)}}
\usage{
gx.upstreamAnalysisTransfacGeneWays(inputYesSet, inputNoSet, profile,
  species = "Human (Homo sapiens)", promoterStart = -1000,
  promoterEnd = 100, resultFolder, skipCompleted = T, wait = T,
  verbose = F)
}
\arguments{
\item{inputYesSet}{yes/positive/foreground gene set}

\item{inputNoSet}{no/negative/background gene set}

\item{profile}{matrix profile}

\item{species}{species of the input track genome}

\item{promoterStart}{first base of a promoter relative to TSS}

\item{promoterEnd}{last base of a promoter relative to TSS}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\value{
the job id of the submitted task. The job id can be used to retrieve information about the status of the analysis.
}
\description{
Runs the workflow \emph{Upstream analysis (TRANSFAC(R) and GeneWays)}
}
\keyword{GeneWays}
\keyword{TRANSFAC,}
\keyword{analysis,}
\keyword{upstream}
\keyword{workflow,}
