% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.vennDiagrams}
\alias{gx.vennDiagrams}
\title{Create a Venn diagram for up to three tables.}
\usage{
gx.vennDiagrams(table1Path, table2Path, table3Path = "",
  table1Name = "Left", table2Name = "Right", table3Name = "Middle",
  output, simple = T, wait = T, verbose = F)
}
\arguments{
\item{table1Path}{path to first (left) table}

\item{table2Path}{path to second (right) table}

\item{table3Path}{path to third (middle) table (optional)}

\item{table1Name}{name/title of left table}

\item{table2Name}{name/title of right table}

\item{table3Name}{name/title of middle table}

\item{output}{output path (does not need to exist)}

\item{simple}{set true for circle diameters proportional to group size}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
Create a Venn diagram for up to three tables.
}
\keyword{Venn,}
\keyword{diagram}
