% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.analysis.list}
\alias{gx.analysis.list}
\title{Lists available analysis tools}
\usage{
gx.analysis.list()
}
\description{
Wraps rbiouml function biouml.analysis.list
}
\seealso{
\code{\link[rbiouml]{biouml.analysis.list}}
}
\keyword{analysis,}
\keyword{list}
