% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.analysis.parameters}
\alias{gx.analysis.parameters}
\title{Shows parameters for the specified analysis tool}
\usage{
gx.analysis.parameters(analysisName)
}
\arguments{
\item{analysisName}{name of the analysis tool}
}
\value{
a list with parameter names and descriptions
}
\description{
Returns a list with parameter names and descriptions.
}
\seealso{
\code{\link{gx.analysis}}

\code{\link{gx.analysis.list}}

\code{\link[rbiouml]{biouml.analysis.parameters}}
}
\keyword{analysis,}
\keyword{parameters}
