% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.analysis}
\alias{gx.analysis}
\title{Execute analysis using one of the integrated tools}
\usage{
gx.analysis(analysisName, parameters = list(), wait = T, verbose = T)
}
\arguments{
\item{analysisName}{name of analysis tool}

\item{parameters}{parameters to configure the analysis}

\item{wait}{TRUE to wait for the analysis to complete. Default: TRUE}

\item{verbose}{TRUE to see some progress info. Default: TRUE}
}
\description{
Wraps rbiouml function biouml.analysis
}
\examples{
gx.analysis()

}
\seealso{
\code{\link{rbiouml::biouml.analysis}}
}
\keyword{analysis}
