% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.createProject}
\alias{gx.createProject}
\title{Create a new project}
\usage{
gx.createProject(name, description = "New platform project")
}
\arguments{
\item{name}{name of the project}

\item{description}{a short description of the project}
}
\value{
a string containing the status of the request in JSON format
}
\description{
This function creates a new project within the platform workspace.
A project is a prerequisite to be able to upload data, perform analyses or
to run workflows.
}
