% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.delete}
\alias{gx.delete}
\title{Delete a workspace item}
\usage{
gx.delete(folder, name)
}
\arguments{
\item{folder}{folder that contains the item}

\item{name}{name of the item}
}
\description{
This function deletes an item within the workspace. The item may be
a folder or another type of element.
}
\details{
Handle with care, because deletion is irreversible without a suitable
backup routine in place.
}
