% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.delete}
\alias{gx.delete}
\title{Delete a workspace item}
\usage{
gx.delete(folder, name)
}
\arguments{
\item{folder}{folder that contains the item}

\item{name}{name of the item}
}
\description{
Delete a workspace item
}
\examples{

# Note that the folder and name probably require adaptation

gx.delete("example-folder","example-name")

}
