% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.ebarraysWorkflow}
\alias{gx.ebarraysWorkflow}
\title{Runs the workflow \emph{Compute differentially expressed genes using EBarrays}}
\usage{
gx.ebarraysWorkflow(inputTable, probeType = "Probes: Affymetrix",
  species = "Human (Homo sapiens)", controlName = "Control",
  controlColumns = c(), conditions = list(), resultFolder,
  skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with expression data}

\item{probeType}{type of probes}

\item{species}{species of the input track genome}

\item{controlName}{name for the control column group}

\item{controlColumns}{list of control column names}

\item{conditions}{a list with condition names and columns}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
Runs the workflow \emph{Compute differentially expressed genes using EBarrays}
}
\keyword{EBarrays}
\keyword{differential}
\keyword{expression,}
\keyword{workflow,}
