% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.export.parameters}
\alias{gx.export.parameters}
\title{Returns parameters defined for given export and path to export from}
\usage{
gx.export.parameters(path, exporter)
}
\description{
Wraps rbiouml function biouml.exporter.parameters
}
\examples{
gx.exporter.parameters("example-path","example-exporter")

}
\seealso{
\code{\link{rbiouml::biouml.exporter.parameters}}
}
\keyword{exporter,}
\keyword{parameters}
