% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.export.parameters}
\alias{gx.export.parameters}
\title{Returns parameters defined for given export and path to export from}
\usage{
gx.export.parameters(path, exporter)
}
\arguments{
\item{path}{path of the object to export}

\item{exporter}{exporter to use for the export}
}
\description{
Given the path of an item to export, this function shows the
parameters for a selected exporter.
}
\seealso{
\code{\link{gx.export}}

\code{\link{gx.exporters}}

\code{\link[rbiouml]{biouml.export.parameters}}
}
\keyword{exporter,}
\keyword{parameters}
