% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.export}
\alias{gx.export}
\title{Exports item using specified exporter}
\usage{
gx.export(path, exporter = "Tab-separated text (*.txt)",
  exporter.params = list(), target.file = "genexplain.out")
}
\arguments{
\item{path}{platform path of item to export}

\item{exporter}{exporter to use for export}

\item{exporter.params}{parameters of the exporter}

\item{target.file}{local file to export to}
}
\value{
\code{NULL}
}
\description{
This function exports a specified item from the platform workspace
to a local file.
The function \emph{gx.exporters} lists available exporters and the
parameters of an exporter can be inspected using the function \emph{gx.export.parameters}.
}
\seealso{
\code{\link{gx.exporters}}

\code{\link{gx.export.parameters}}

\code{\link[rbiouml]{biouml.export}}
}
\keyword{export}
