% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.exporters}
\alias{gx.exporters}
\title{Lists available exporters}
\usage{
gx.exporters()
}
\description{
Wraps rbiouml function biouml.exporters
}
\examples{
gx.exporters()

}
\seealso{
\code{\link{rbiouml::biouml.exporters}}
}
\keyword{exporter,}
\keyword{list}
