% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.import}
\alias{gx.import}
\title{Imports a file into the platform}
\usage{
gx.import(file, parentPath, importer, importer.params = list())
}
\arguments{
\item{file}{local file to import}

\item{parentPath}{path of folder to import into}

\item{importer}{importer to use}

\item{importer.params}{parameters for specified importer}
}
\value{
the platform path of the imported item
}
\description{
This function uploads a file into the platform using 
the specified importer configured by its
parameters.
The function \emph{gx.importers} lists the available
importers. Parameters for a specific importer can
be inspected using the \emph{gx.import.parameters} function.
}
\seealso{
\code{\link{gx.importers}}

\code{\link{gx.import.parameters}}

\code{\link[rbiouml]{biouml.import}}
}
\keyword{import}
