% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.import}
\alias{gx.import}
\title{Imports a file into the platform}
\usage{
gx.import(file, parentPath, importer, importer.params = list())
}
\arguments{
\item{file}{local file to import}

\item{parentPath}{path of folder to import into}

\item{importer}{importer to use}

\item{importer.params}{parameters for specified importer}
}
\description{
Wraps rbiouml function biouml.import
}
\seealso{
\code{\link[rbiouml]{biouml.import}}
}
\keyword{import}
