% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.importers}
\alias{gx.importers}
\title{Lists available importers}
\usage{
gx.importers()
}
\description{
Returns a vector with available importer names. An importer
uploads a file to the platform, where it is imported as a
certain data type, e.g. as pathway or molecular network.
}
\seealso{
\code{\link{gx.import}}

\code{\link{gx.import.parameters}}

\code{\link[rbiouml]{biouml.importers}}
}
\keyword{importer,}
\keyword{list}
