% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.job.info}
\alias{gx.job.info}
\title{Gets status info for a job running on the platform}
\usage{
gx.job.info(jobID)
}
\description{
Wraps rbiouml function biouml.job.info
}
\seealso{
\code{\link[rbiouml]{biouml.job.info}}
}
\keyword{info}
\keyword{job,}
