% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform-tools.R
\name{gx.limmaWorkflow}
\alias{gx.limmaWorkflow}
\title{Runs the workflow \emph{Compute differentially expressed genes using Limma}}
\usage{
gx.limmaWorkflow(inputTable, probeType = "Probes: Affymetrix",
  species = "Human (Homo sapiens)", conditions = list(), resultFolder,
  skipCompleted = T, wait = T, verbose = F)
}
\arguments{
\item{inputTable}{input table with expression data}

\item{probeType}{type of probes}

\item{species}{species of the input track genome}

\item{conditions}{a list with condition names and columns}

\item{resultFolder}{path of result folder}

\item{skipCompleted}{skip already completed steps}

\item{wait}{set true to wait for the analysis to complete}

\item{verbose}{set true for more progress info}
}
\description{
Runs the workflow \emph{Compute differentially expressed genes using Limma}
}
\keyword{analysis}
\keyword{upstream}
\keyword{workflow,}
