% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.login}
\alias{gx.login}
\title{Signs into geneXplain platform}
\usage{
gx.login(server = "https://platform.genexplain.com", user = "",
  password = "")
}
\arguments{
\item{server}{server to connect to}

\item{user}{username}

\item{password}{password}
}
\description{
Wraps rbiouml function biouml.login
}
\seealso{
\code{\link[rbiouml]{biouml.login}}
}
\keyword{login}
