% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.ls}
\alias{gx.ls}
\title{List contents of platform folder}
\usage{
gx.ls(path, extended = F)
}
\arguments{
\item{path}{path of folder to list}

\item{extended}{get extended info including data types and whether folders have children}
}
\description{
This function returns a list of contents of specified folder.
Project folders are usually located under \emph{data/Projects}.
The \emph{extended} output contains in addition an indicator whether
a folder element has further child elements and the type of the element.
}
\seealso{
\code{\link[rbiouml]{biouml.ls}}
}
\keyword{ls}
