% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.put}
\alias{gx.put}
\title{Uploads a table to the platform}
\usage{
gx.put(path, value)
}
\arguments{
\item{path}{platform path of new table}

\item{value}{R object to put into platform}
}
\description{
This function stores a data frame in the specified path
of the platform workspace.
}
\seealso{
\code{\link[rbiouml]{biouml.put}}
}
\keyword{put}
