% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXplain.R
\name{gx.workflow}
\alias{gx.workflow}
\title{Executes workflow with specified parameters}
\usage{
gx.workflow(path, parameters = list(), wait = T, verbose = T)
}
\description{
Wraps rbiouml function biouml.workflow
}
\examples{
gx.workflow("path/to/example/workflow", list("param1"="value1"))

}
\seealso{
\code{\link{rbiouml::biouml.workflow}}
}
\keyword{workflow}
